+++ parser/src/lexer.spec.ts
@@ -971,6 +971,200 @@ hello
     ]);
   });
 
+  describe('match', () => {
+    it('match with branches', () => {
+      const tokens = tokenize(`
+{ match this_is_a_variable
+  'value_a':
+    This is a line
+  'value_b':
+    This is another line
+  default:
+    This is the default line
+}`).getAll();
+
+      expect(tokens).toEqual([
+        { token: TOKENS.LINE_BREAK, line: 1, column: 0 },
+        { token: TOKENS.BRACE_OPEN, line: 1, column: 0, },
+        { token: TOKENS.KEYWORD_MATCH, line: 1, column: 2, },
+        { token: TOKENS.IDENTIFIER, value: 'this_is_a_variable', line: 1, column: 8, },
+        { token: TOKENS.INDENT, line: 2, column: 0 },
+        { token: TOKENS.STRING_LITERAL, value: 'value_a', line: 2, column: 2, },
+        // { token: TOKENS.COLON, line: 2, column: 9, },
+        { token: TOKENS.INDENT, line: 3, column: 2 },
+        {
+          token: TOKENS.TEXT,
+          value: 'This is a line',
+          line: 3,
+          column: 4,
+        },
+        { token: TOKENS.DEDENT, line: 4, column: 2 },
+        { token: TOKENS.STRING_LITERAL, value: 'value_b', line: 4, column: 2, },
+        // { token: TOKENS.COLON, line: 4, column: 9, },
+        { token: TOKENS.INDENT, line: 5, column: 2 },
+        {
+          token: TOKENS.TEXT,
+          value: 'This is another line',
+          line: 5,
+          column: 4,
+        },
+        { token: TOKENS.DEDENT, line: 6, column: 2 },
+        { token: TOKENS.KEYWORD_DEFAULT, line: 6, column: 2, },
+        // { token: TOKENS.COLON, line: 6, column: 6, },
+        { token: TOKENS.INDENT, line: 7, column: 2 },
+        {
+          token: TOKENS.TEXT,
+          value: 'This is the default line',
+          line: 7,
+          column: 4,
+        },
+        { token: TOKENS.DEDENT, line: 8, column: 2 },
+        { token: TOKENS.DEDENT, line: 8, column: 0 },
+        { token: TOKENS.BRACE_CLOSE, line: 8, column: 0, },
+        { token: TOKENS.EOF, line: 8, column: 1 },
+      ]);
+    });
+
+    it('match complex statement', () => {
+      const tokens = tokenize(`
+{ match this_is_a_variable && this_other > 10
+  'value_a':
+    This is a line
+}`).getAll();
+
+      expect(tokens).toEqual([
+        { token: TOKENS.LINE_BREAK, line: 1, column: 0 },
+        { token: TOKENS.BRACE_OPEN, line: 1, column: 0, },
+        { token: TOKENS.KEYWORD_MATCH, line: 1, column: 2, },
+        { token: TOKENS.IDENTIFIER, value: 'this_is_a_variable', line: 1, column: 8, },
+        { token: TOKENS.AND, line: 1, column: 27, },
+        { token: TOKENS.IDENTIFIER, value: 'this_other', line: 1, column: 30, },
+        { token: TOKENS.GREATER, line: 1, column: 41, },
+        { token: TOKENS.NUMBER_LITERAL, value: '10', line: 1, column: 43, },
+        { token: TOKENS.INDENT, line: 2, column: 0 },
+        { token: TOKENS.STRING_LITERAL, value: 'value_a', line: 2, column: 2, },
+        { token: TOKENS.INDENT, line: 3, column: 2 },
+        {
+          token: TOKENS.TEXT,
+          value: 'This is a line',
+          line: 3,
+          column: 4,
+        },
+        { token: TOKENS.DEDENT, line: 4, column: 2 },
+        { token: TOKENS.DEDENT, line: 4, column: 0 },
+        { token: TOKENS.BRACE_CLOSE, line: 4, column: 0, },
+        { token: TOKENS.EOF, line: 4, column: 1 },
+      ]);
+    });
+
+    it('branches in the same line', () => {
+      const tokens = tokenize(`
+{ match this_is_a_variable
+  true: This is a line
+  default: This is the default line
+}`).getAll();
+
+      expect(tokens).toEqual([
+        { token: TOKENS.LINE_BREAK, line: 1, column: 0 },
+        { token: TOKENS.BRACE_OPEN, line: 1, column: 0, },
+        { token: TOKENS.KEYWORD_MATCH, line: 1, column: 2, },
+        { token: TOKENS.IDENTIFIER, value: 'this_is_a_variable', line: 1, column: 8, },
+        { token: TOKENS.INDENT, line: 2, column: 0 },
+        { token: TOKENS.BOOLEAN_LITERAL, value: 'true', line: 2, column: 2, },
+        {
+          token: TOKENS.TEXT,
+          value: 'This is a line',
+          line: 2,
+          column: 8,
+        },
+        { token: TOKENS.KEYWORD_DEFAULT, line: 3, column: 2, },
+        {
+          token: TOKENS.TEXT,
+          value: 'This is the default line',
+          line: 3,
+          column: 11,
+        },
+        { token: TOKENS.DEDENT, line: 4, column: 0 },
+        { token: TOKENS.BRACE_CLOSE, line: 4, column: 0, },
+        { token: TOKENS.EOF, line: 4, column: 1 },
+      ]);
+    });
+
+    it('mixed with content', () => {
+      const tokens = tokenize(`
+this is a line outside the match
+{ match this_is_a_variable
+  'value_a':
+    This is a line
+}
+this is a line after the match
+`).getAll();
+
+      expect(tokens).toEqual([
+        { token: TOKENS.TEXT, value: 'this is a line outside the match', line: 1, column: 0 },
+        { token: TOKENS.LINE_BREAK, line: 2, column: 0 },
+        { token: TOKENS.BRACE_OPEN, line: 2, column: 0, },
+        { token: TOKENS.KEYWORD_MATCH, line: 2, column: 2, },
+        { token: TOKENS.IDENTIFIER, value: 'this_is_a_variable', line: 2, column: 8, },
+        { token: TOKENS.INDENT, line: 3, column: 0 },
+        { token: TOKENS.STRING_LITERAL, value: 'value_a', line: 3, column: 2, },
+        { token: TOKENS.INDENT, line: 4, column: 2 },
+        {
+          token: TOKENS.TEXT,
+          value: 'This is a line',
+          line: 4,
+          column: 4,
+        },
+        { token: TOKENS.DEDENT, line: 5, column: 2 },
+        { token: TOKENS.DEDENT, line: 5, column: 0 },
+        { token: TOKENS.BRACE_CLOSE, line: 5, column: 0, },
+        { token: TOKENS.TEXT, value: 'this is a line after the match', line: 6, column: 0 },
+        { token: TOKENS.EOF, line: 7, column: 0 },
+      ]);
+    });
+
+    it('handles error scenario gracefully', () => {
+      const tokens = tokenize(`
+{ match
+  true:
+    This is a line
+}
+
+{
+  match this_is_a_variable
+     This
+}
+`).getAll();
+
+      const relevantTokens = tokens.slice(0, 15);
+
+      expect(relevantTokens).toEqual([
+        { token: TOKENS.LINE_BREAK, line: 1, column: 0 },
+        { token: TOKENS.BRACE_OPEN, line: 1, column: 0, },
+        { token: TOKENS.KEYWORD_MATCH, line: 1, column: 2, },
+        { token: TOKENS.INDENT, line: 2, column: 0 },
+        { token: TOKENS.BOOLEAN_LITERAL, value: 'true', line: 2, column: 2, },
+        { token: TOKENS.INDENT, line: 3, column: 2 },
+        {
+          token: TOKENS.TEXT,
+          value: 'This is a line',
+          line: 3,
+          column: 4,
+        },
+        { token: TOKENS.DEDENT, line: 4, column: 2 },
+        { token: TOKENS.DEDENT, line: 4, column: 0 },
+        { token: TOKENS.BRACE_CLOSE, line: 4, column: 0, },
+
+        { token: TOKENS.LINE_BREAK, line: 6, column: 0 },
+        { token: TOKENS.BRACE_OPEN, line: 6, column: 0, },
+        { token: TOKENS.KEYWORD_MATCH, line: 7, column: 2, },
+        { token: TOKENS.IDENTIFIER, value: 'this_is_a_variable', line: 7, column: 8, },
+
+        { token: TOKENS.INDENT, line: 8, column: 0 },
+      ]);
+    });
+  });
+
   it('events trigger', () => {
     const tokens = tokenize(`
 { trigger event_name }