+++ parser/src/lexer.ts
@@ -41,6 +41,8 @@ export const TOKENS = {
   KEYWORD_SET: 'set',
   KEYWORD_TRIGGER: 'trigger',
   KEYWORD_WHEN: 'when',
+  KEYWORD_MATCH: 'match',
+  KEYWORD_DEFAULT: 'default',
   ASSIGN: '=',
   ASSIGN_SUM: '+=',
   ASSIGN_SUB: '-=',
@@ -60,6 +62,8 @@ enum LexerMode {
   QSTRING,
   LOGIC,
   VARIATIONS,
+  MATCH,
+  MATCH_BODY,
 };
 
 const tokenFriendlyHint = {
@@ -146,6 +150,9 @@ export function tokenize(input: string): TokenList {
     }
   };
 
+  const matchBodyIndent: number[] = [];
+  let isInlineMatchBranch: boolean = false;
+
   const isCurrentMode = (mode: LexerMode): boolean => {
     return modes[0] === mode;
   };
@@ -168,6 +175,11 @@ export function tokenize(input: string): TokenList {
         const previousIndent = indent[0];
         column += indentation;
         indent.unshift(indentation);
+
+        if (isCurrentMode(LexerMode.MATCH_BODY)) {
+          matchBodyIndent.push(previousIndent);
+        }
+
         return { token: TOKENS.INDENT, line: initialLine, column: previousIndent };
     }
 
@@ -181,10 +193,20 @@ export function tokenize(input: string): TokenList {
         indent.shift();
         column = indent[0];
         tokens.push({ token: TOKENS.DEDENT, line, column });
+
+        if (isMatchBodyIndentLevel(column)) {
+          popMode();
+          matchBodyIndent.shift();
+        }
     }
 
     return tokens;
   };
+
+  const isMatchBodyIndentLevel = (currentIndent: number): boolean => (
+    isCurrentMode(LexerMode.MATCH_BODY) &&
+    matchBodyIndent[matchBodyIndent.length - 1] === currentIndent
+  );
 
   // handle comments
   const handleComments = (): void => {
@@ -204,6 +226,11 @@ export function tokenize(input: string): TokenList {
     }
 
     if (isCurrentMode(LexerMode.OPTION)) {
+      popMode();
+      return;
+    }
+
+    if (isCurrentMode(LexerMode.MATCH_BODY) && isInlineMatchBranch) {
       popMode();
     }
   };
@@ -548,8 +575,79 @@ export function tokenize(input: string): TokenList {
     return { token: TOKENS.MINUS, line, column: initialColumn };
   };
 
+  const handleLogicOrMatchBlockStart = (): TokenHandlerReturn => {
+    const blockStart = handleLogicBlockStart();
 
-  const handleLogicBlockStart = () => {
+    while (input[position] && input[position].match(/[ \n\t]/)) {
+      if (input[position] === '\n') {
+        line += 1;
+        column = 0;
+      } else {
+        column += 1;
+      }
+      position += 1;
+    }
+
+    if (checkSequence(input, position, 'match')) {
+      const token = { token: TOKENS.KEYWORD_MATCH, line, column };
+      position += 5;
+      column += 5;
+
+      handleSpace();
+
+      const statement = handleLogicStatement();
+
+      stackMode(LexerMode.MATCH);
+
+      if (!statement) {
+        return [].concat(blockStart, token);
+      }
+
+      return [].concat(blockStart, token, statement);
+    }
+
+    return blockStart;
+  }
+
+  const handleMatchBlock = (): TokenHandlerReturn | undefined => {
+    if (input[position] === "}") {
+      column += 1;
+      position += 1;
+      popMode(); // pop match
+      popMode(); // pop logic
+      return { token: TOKENS.BRACE_CLOSE, line, column: column - 1 };
+    }
+
+    let statement: TokenHandlerReturn | undefined;
+
+    if (checkSequence(input, position, 'default:')) {
+      statement = { token: TOKENS.KEYWORD_DEFAULT, line, column: column };
+      position += 7;
+      column += 7;
+    } else {
+      statement = handleLogicStatement();
+    }
+
+    if (input[position] === ":") {
+      column += 1;
+      position += 1;
+      stackMode(LexerMode.MATCH_BODY);
+
+      isInlineMatchBranch = false;
+      let p = position
+      while(input[p] && input[p] != '\n') {
+        if (input[p] != ' ') {
+          isInlineMatchBranch = true;
+          break;
+        }
+        p += 1;
+      }
+    }
+
+    return statement;
+  };
+
+  const handleLogicBlockStart = (): TokenHandlerReturn => {
     const initialColumn = column;
     column += 1;
     position += 1;
@@ -577,7 +675,7 @@ export function tokenize(input: string): TokenList {
 
   const keywords = [
     'is', 'isnt', 'or', 'and', 'not', 'true', 'false', 'null',
-    'set', 'trigger', 'when'
+    'set', 'trigger', 'when', 'match',
   ];
 
   const handleLogicIdentifier = (): TokenHandlerReturn => {
@@ -627,7 +725,6 @@ export function tokenize(input: string): TokenList {
       case 'when':
         return { token: TOKENS.KEYWORD_WHEN, line, column: initialColumn };
     }
-
   };
 
   const handleLogicNot = (): Token => {
@@ -679,15 +776,19 @@ export function tokenize(input: string): TokenList {
   };
 
   const handleLogicBlock = (): TokenHandlerReturn => {
+    if (input[position] === '}') {
+      return handleLogicBlockStop();
+    }
+
+    return handleLogicStatement();
+  };
+
+  const handleLogicStatement = (): TokenHandlerReturn => {
     if (input[position] === '"' || input[position] == "'") {
       if (currentQuote === null) {
         currentQuote = input[position]
       }
       return handleLogicString();
-    }
-
-    if (input[position] === '}') {
-      return handleLogicBlockStop();
     }
 
     if (input[position] === '(') {
@@ -842,11 +943,19 @@ export function tokenize(input: string): TokenList {
     }
 
     if (!isCurrentMode(LexerMode.QSTRING) && input[position] === '{') {
-      return handleLogicBlockStart();
+      return handleLogicOrMatchBlockStart();
     }
 
     if(isCurrentMode(LexerMode.LOGIC)) {
       const response = handleLogicBlock();
+
+      if (response)  {
+        return response;
+      }
+    }
+
+    if(isCurrentMode(LexerMode.MATCH)) {
+      const response = handleMatchBlock();
 
       if (response)  {
         return response;