+++ cli/CHANGELOG.md
@@ -2,12 +2,11 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
-## Unreleased
+## 3.2.0 (2024-11-26)
 
 ### Changed
 
 - Support events with parameters (e.g `{ trigger event(param1, param2, "text", ...) }`)
-
 
 ## 3.1.0 (2024-10-07)