+++ interpreter/CHANGELOG.md
@@ -4,10 +4,15 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Unreleased
 
+### Added
+
+- Match block support
+
+## 4.2.0 (2024-11-26)
+
 ### Changed
 
 - Support events with parameters (e.g `{ trigger event(param1, param2, "text", ...) }`)
-
 
 ## 4.1.0 (2024-10-07)