+++ parser/CHANGELOG.md
@@ -4,6 +4,12 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Unreleased
 
+### Added
+
+- Add match block support
+
+## 2.5.0 (2024-11-26)
+
 ### Changed
 
 - Support events with parameters (e.g `{ trigger event(param1, param2, "text", ...) }`)